+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 21.2.1 [#613](https://github.com/openfisca/openfisca-core/pull/613)
 
-- Fix a bug introduced in `21.0.2`:
-  - The result of a formula returning an Enum value was not properly encoded
+- Fix two bugs that appeared with 21.2.0:
+  - Properly encode the result of a formula returning an Enum value
+  - Enable storing an Enum value on disk
 
 ## 21.2.0 [#601](https://github.com/openfisca/openfisca-core/pull/601)